Move contents of SiPixelTemplateDBObjectESProducer header to source file
